requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==22.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Django==4.1
djangorestframework==3.14.0
drf-spectacular==0.25.1
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
pyrsistent==0.19.3
pytz==2022.7
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
